Split huge unbrake.go file into multiple files
Resolves #248
